Ahaha I forgot how much you Ruby nerds love your convention over your configuration
... personally I have no love for either; as a child of Steve Jobs I build things to be
maximally sedulous, instead of needy (requiring configuration, postmodern) or assholic
(demanding convention, modern) -- which is a lot more work, in both depth and breath, than
writing a half-assed configuration file DSL, or a blog post that frames your softwares' crufty
warts as a bold declaration of identity and persona(*)... Ergh. Fuck that false dichotomy, fuck config
files and their maddening inconsistencies, fuck boilerplate, fuck all that shit.

*) e.g. the point of this commit, which is that apparently homebrew tap repos need to start
with 'homebrew-' ... which I get that, it keeps them from having to maintain a central
list of taps by allowing tap repo publishers to opt into having their repos tapped by using
this name convention, which is easier than (say) having tap repos contain certain files or
metadata, and certainly it is easier than doing a central-list thing which is always a pain
at best, etc etc. I totally get that -- it's just that it's not really documented as such
outside the code -- I have skimmed homebrew's codebase once but not read it extensively,
which you have to say that's, like, more regard than most non-Rubyist non-Homebrew-team people
probably generally give, rite??... Ergh. Yeah. I mean so yeah, first off, it wass kind of
dumb that I had to rename the repo after I created it, which makes me look retarded
on GitHub. And now it is secondarily, persistently dumb that I have this public repo
called 'homebrew-brew' rattling around. I could change it to something less immediately
inane, sure, but whatever I do call it it is going to be 'homebrew-' plus whatever that
is. The whole thing is just over the line, in terms of how stupid it is (such that I have
been miffed into acting like a buffoon, as this commit note illustrates) -- but it's
not anywhere near as egregiously stupid as it'd need to be to get any sort of critical-
mass-type attention from the committers (who understandably have to have very high
bullshit thresholds in their interactions with their dev community, considering their
project is a super-popular piece of pivotal -- to some, critical -- tech infrastructure).
BLECH. In a nutshell.